Addressed comments after review
