+++ src/redux/slices/drag-mode-slice.ts
