+++ src/redux/slices/uv-debug-texture-slice.ts
